+++ cancer_api/__init__.py
@@ -4,13 +4,13 @@ __init__.py
 Where it all begins.
 """
 
-__version__ = "0.1.5"
-
 from main import connect, get_connection, set_connection
 from connections import MysqlConnection, SqliteConnection
 from mutations import SingleNucleotideVariant, Indel, CopyNumberVariation, StructuralVariation
 from metadata import Patient, Sample, Library
 from annotations import Gene, Transcript, Exon, Protein
-import files
-import parsers
-import utils
+from files import *
+from parsers import *
+from utils import *
+
+__version__ = "0.2.0"